placed comoponents into their own files
